requirements.txt
openai>=0.27.0
numpy>=1.21.0
matplotlib>=3.4.0
Pillow>=8.0.0 